api: Initial API design
This commit creates the `litev.h` file which contains the C API for the
litev library.

The hearth of the event loop is `struct litev_base` which is being
created by the `litev_init()` and deleted by the `litev_free()`
functions.

The event loop itself is launched by the `litev_dispatch()` and stopped
by the `litev_break()` functions.

An event itself is of type `struct litev_ev` and consists of the file
descriptor, the condition (read | write), the callback function to be
called when the condition is satisfied and an arbitrary pointer of user
data which is the last argument of the callback function.

An event is identified by the file descriptor/condition pair and must
be unique within an event loop.

The events are then being added and removed to and from the event loop
by the `litev_add()` and the `litev_del()` functions.