api: Implement the internal API
This commit introduces the file `litev-internal.h` which holds the
internal API for litev, which is going to be used to provide a common
interface for the various backends of the kernel event notification
APIs.

The API for the kernel event notification API backends demands the
following functions:
- `init()`, which initializes the `EV_API_DATA` and generally prepares
  the kernel notification API.
- `free()`, which obviously frees the resources associated with the
  kernel event notification API.
- `poll()`, which is called regularly by `litev_dispatch()` until the
  event loop is stopped.  This function waits until some condition on
  a file descriptor has been triggered and then executes the
  appropriate callbacks.
- `add()` and `del()`, which obviously add/remove events from the
  kernel event notification API.

Beside this, this commit also adds the declaration for
`struct litev_base`.