api: Change the prototype of the `poll` function
Previously the only argument that the `poll()` function from
`struct litev_ev_api` accepted was of type `EV_API_DATA *`.
However this is not sufficient because the `poll()` function eventually
calls the user provided callback which needs to have the
`struct litev_base *`.

Because of this, this commit changes the argument type from
`EV_API_DATA *` to `struct litev_base *`, making it the only function
inside `struct litev_ev_api` to have full access to the event loop
data.