kqueue: Add code for a doubly linked list
Because it is nasty for the caller process to keep its copy of
`struct litev_ev` around, litev makes a copy of it on the heap.

However, `kqueue(2)` offers no way of obtaining all registered events,
making it impossible to perform a clean-up without memory leaks.

To get around this issue, we will not store the pointer to the heap
copy of the `struct litev_ev` inside `kqueue(2)`, instead we store
a pointer to a node of a doubly linked list.

Because the copies are now stored in a doubly linked list, we can
iterate over something during the clean-up and have something
that is inexpensive in terms of adding/removing nodes dynamically.

A deeper motivation for this can be found inside the comment above
the declaration of `struct udata_node` inside `kqueue.c`.