api: Remove `struct litev_base *` from the cb
This commit removes the `struct litev_base *` as an argument from the
function prototype of the `struct litev_ev` callback.

This is done because it may not always be necessary to have access to
the event loop and in the case it is still desired, the caller is free
to use `udata` or a global variable for it.