kqueue: Implement hash map
Because our original intention to store the copies of `struct litev_ev`
into a doubly linked list turned out to be a bad design decision, we
now use a hash map instead.

This commit adds the implementation for a simple hash map which
identifies the events by the `fd` and the `condition` field.