litev: Actually set `is_dispatched`
Previously the `litev_dispatch()` function did not set the
`is_dispatched` boolean inside `struct litev_base`.  This resulted in
`litev_break()` not being able to stop the event loop as it demands
that this boolean is set to true in order work.

It was spotted while trying to examine the `tcp-echo.c` example.