api: Implement `litev_close()`
This commit adds and implements a new function to the public API called
`litev_close()`, which removes all registered events from the kernel
event notification API associated with that file descriptor and
performs a `close(2)` on it.

Previously, the calling process had to call `litev_del()` before
passing it to `close(2)`.

The reason why this function is added, although the same goal can be
achieved with other ways as well, is because it is a very common
operation and might actually result in a performance increase depending
on the kernel event notification API that is in use.