kqueue: Only compile if `USE_KQUEUE` is defined
Previously, the `kqueue.c` file was always compiled, regardless if the
`kqueue(2)` kernel event notification API is being used or not.

This commit adds the entire `kqueue.c` code base (except for
the preprocessor stuff) into an `#ifdef USE_KQUEUE`, because support
for other kernel event notification APIs is going to be added soon.