poll: Begin work on the `poll(2)` backend
This commit introduces the kernel event notification API detection
inside `litev-internal.h` for the `poll(2)` event notification API,
in case that no better alternative was found (due to the fact that
`poll(2)` is part of POSIX).

Beside this, it also implements the `ev_api_poll()` function, which
however only initializes the `struct litev_ev_api` fields to NULL.
Logic is going to be added within the next commits.