build: Move kernel notification API detection
This commit moves the detection for the kernel notification API that is
going to be used from `litev-internal.h` to its own dedicated file,
`config.h`.

The reason for this is, that the current build fails on non-BSD systems
because they try to include `<sys/event.h>` although it does not exist,
because the inclusion of `litev-internal.h` happens later, though it is
not aware of constants such as `USE_KQUEUE` during that point of time.

With this commit, all source files include `config.h` as their first
file, even before `<sys/types.h>`.  Directly after this, they ensure
that this only builds, if the appropriate kernel notification API is
being used, by performing the check such as `#ifdef USE_KQUEUE`.