epoll: Begin work on the `epoll(2)` backend
This commit introduces the kernel event notification API detection
inside `config.h` for the `epoll(2)` kernel event notification API,
in case that the Linux kernel is being used.

Beside this, it also implements the `ev_api_epoll()` function, which
however only initializes the `struct litev_ev_api` fields to NULL.
Logic is going to be added within the next commits.