poll: Do not block if no events are registered
Previously, the `poll_poll()` function blocked independently, no matter
if events were registered or not.  This behavior is unacceptable, as it
makes the adding and removing of events impossible.

This commit introduces a new variable, `nactive_ev`, into the
`struct poll_data`, which has a meaningless value, that only matters if
it is zero or not.  This variable is then being checked by
`poll_poll()` during the very beginning and will cause the function
to return immediately with `LITEV_OK`, if its value is zero.

Now, `poll_poll()` behaves like its `kqueue(2)` and `epoll(2)` pedant:
Return immediately if no events are registered.