perf: Begin work on a performance tests library
This commit begins the work on the performance test library, which
will contain functions that are the same across all performance tests,
such as the socket creation.

Currently, this library only consists of a single header file called
`perf.h`, which contains the default HTTP/1.1 response inside the CPP
definition `PERF_REPLY`.  Future functionality is going to be added
within the next commits.