kqueue: Fix memory leak in `kqueue_close()`
Previously, litev leaked memory while using the `kqueue(3)` kernel
event notification API.  This was because litev did not reduced
the amount of active events, when `litev_close()` was used.

This commit fixes this bug, by using `kqueue_del()` instead of
`hash_del()` for removing the events prior to closing the file
descriptor.